+++ src/com/ludumdare40/com/entities/Creature.java
@@ -3,18 +3,18 @@ package com.ludumdare40.com.entities;
 import org.newdawn.slick.Image;
 
 public abstract class Creature extends Entity {
-	private int health;
+	private float health;
 	
-	public Creature(float x, float y, Image img, int health) {
+	public Creature(float x, float y, Image img, float health) {
 		super(x, y, img);
 		this.health = health;
 	}
 
-	public int getHealth() {
+	public float getHealth() {
 		return health;
 	}
 
-	public void hurt(int damage) {
+	public void hurt(float damage) {
 		health -= damage;
 	}
