+++ src/com/ludumdare40/com/entities/Person.java
@@ -3,7 +3,7 @@ package com.ludumdare40.com.entities;
 import org.newdawn.slick.Image;
 
 public class Person extends Creature{
-    public Person(float x, float y, Image img, int health) {
+    public Person(float x, float y, Image img, float health) {
         super(x, y, img, health);
     }
